fix(build): assemble proper canada site menu
Canada.ca theme is sooo special, it has to have its own site menus,
and not like the other themes, no-no, instead of just generic template that
can be used to generate language-specific ones, they already have language-specific
templates for English and French so we can't use our generic assemble task. Need to override it.

Also, the default template references sitemenu from a different place, so that was also fixed.